requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

deeplake==3.9.9
GitPython==3.1.43
langchain==0.2.1
langchain-community==0.2.4
python-dotenv==1.0.1
streamlit==1.34.0
torch==2.3.1
